multi-agent-systems/news_reporter/requirements-dev.txt
semantic-kernel>=1.21.0
azure-ai-projects>=1.0.0
azure-identity>=1.21.0
python-dotenv>=1.0.0
asyncio
